Chess_Chaos/requirements.txt
python-chess
requests 